Refactor database handlers.
<h1>Rationale:</h1>

This commit moves database handlers to a handler package.  This significantly reduces the size of DatabaseListManager in anticipation of that class growing in size as experiences (games) are added to the database.  No functional, cosmetic or other changes are included with this commit.  Simply moving code from one file to another.
