+++ app/src/main/java/com/pajato/android/gamechat/database/DatabaseRegistrar.java
@@ -23,6 +23,7 @@ import com.google.firebase.database.ChildEventListener;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
+import com.pajato.android.gamechat.database.handler.DatabaseEventHandler;
 
 import java.util.HashMap;
 import java.util.Locale;
